+++ axelor-account/src/main/java/com/axelor/apps/account/service/MoveLineExportServiceImpl.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Axelor Business Solutions
  *
  * Copyright (C) 2018 Axelor (<http://axelor.com>).
@@ -120,6 +120,7 @@ public class MoveLineExportServiceImpl implements MoveLineExportService{
 
 			if (i % 10 == 0) { JPA.clear(); }
 			if (i++ % 100 == 0) { log.debug("Process : {} / {}" , i, moveListSize); }
+
 		}
 
 	}
@@ -913,9 +914,14 @@ public class MoveLineExportServiceImpl implements MoveLineExportService{
 		List<String[]> allMoveLineData = new ArrayList<>();
 		Company company = accountingReport.getCompany();
 		
+		LocalDate interfaceDate = accountingReport.getDate();
+		String exportNumber = this.getSaleExportNumber(company);
+
 		String moveLineQueryStr = "";
 		moveLineQueryStr += String.format(" AND self.move.company = %s", company.getId());
-		moveLineQueryStr += String.format(" AND self.move.period.year = %s", accountingReport.getYear().getId());
+		if(accountingReport.getYear() != null){
+			moveLineQueryStr += String.format(" AND self.move.period.year = %s", accountingReport.getYear().getId());
+		}
 		
 		if(accountingReport.getPeriod() != null)	{
 			moveLineQueryStr += String.format(" AND self.move.period = %s", accountingReport.getPeriod().getId());
@@ -933,15 +939,18 @@ public class MoveLineExportServiceImpl implements MoveLineExportService{
 		}
 
         if(administration){
-            moveLineQueryStr += String.format("AND self.move.ignoreInAccountingOk = false");
+            moveLineQueryStr += String.format(" AND self.move.ignoreInAccountingOk = false");
         }
 
 		List<MoveLine> moveLineList = moveLineRepo.all().filter("self.move.statusSelect = ?1" + moveLineQueryStr, MoveRepository.STATUS_VALIDATED).order("date").order("name").fetch();
-		if(moveLineList.size() > 0) {
-			
+		if(!moveLineList.isEmpty()) {
+			List<Move> moveList = new ArrayList<>();
 			for (MoveLine moveLine : moveLineList) {
-				String items[] = new String[18];
+				String[] items = new String[18];
 				Move move = moveLine.getMove();
+				if(!moveList.contains(move)){
+					moveList.add(move);
+				}
 				Journal journal = move.getJournal();
 				items[0] = journal.getCode();
 				items[1] = journal.getName();
@@ -991,12 +1000,18 @@ public class MoveLineExportServiceImpl implements MoveLineExportService{
 				}
 				allMoveLineData.add(items);
 			}
+
+			if(!administration){
+				this.updateMoveList(moveList, accountingReport, interfaceDate, exportNumber);
+			}
 		}
 
+		accountingReport = accountingReportRepo.find(accountingReport.getId());
+		
 		String fileName = this.setFileName(accountingReport);
 		String filePath = accountConfigService.getExportPath(accountConfigService.getAccountConfig(company));
 		//TODO create a template Helper
-
+		
 		new File(filePath).mkdirs();
 		log.debug("Full path to export : {}{}" , filePath, fileName);
 //		CsvTool.csvWriter(filePath, fileName, '|', null, allMoveLineData);
@@ -1004,12 +1019,13 @@ public class MoveLineExportServiceImpl implements MoveLineExportService{
 		accountingReportRepo.save(accountingReport);
 
 		Path path = Paths.get(filePath+fileName);
-
+		
 		try (InputStream is = new FileInputStream(path.toFile())) {
 			Beans.get(MetaFiles.class).attach(is, fileName, accountingReport);
 		}
 
 	}
+
 
 	/**
 	 * Méthode réalisant l'export SI - Agresso des fichiers détails
@@ -1407,8 +1423,11 @@ public class MoveLineExportServiceImpl implements MoveLineExportService{
 			fileName += accountingReport.getDateTo().format(DateTimeFormatter.ofPattern(DATE_FORMAT_YYYYMMDD));
 		}else if(accountingReport.getPeriod() != null){
 			fileName += accountingReport.getPeriod().getToDate().format(DateTimeFormatter.ofPattern(DATE_FORMAT_YYYYMMDD));
-		}else {
+		}else if(accountingReport.getYear() != null){
 			fileName += accountingReport.getYear().getToDate().format(DateTimeFormatter.ofPattern(DATE_FORMAT_YYYYMMDD));
+		}
+		else {
+			throw new AxelorException(I18n.get(IExceptionMessage.MOVE_LINE_EXPORT_YEAR_OR_PERIOD_OR_DATE_IS_NULL), IException.NO_VALUE);
 		}
 		fileName +=".csv";